requirements.txt
streamlit
requests
openai>=1.40.0
python-dotenv
python-dateutil
beautifulsoup4
lxml
python-docx
sqlmodel
pytz
bcrypt>=4.1.2
streamlit-cookies-manager>=0.2.0
pandas>=2.2.2
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9